refactor: Erase old DatePicke
