--- packages/synapse-bridge/src/patterns/DatePicker/locales.ts
@@ -1,3 +1,0 @@
-export const locales = {
-	openCalendar: 'Ouvrir le calendrier',
-}
