--- packages/synapse-bridge/src/patterns/DatePicker/mixins/maskValue.ts
@@ -1,27 +1,0 @@
-import { defineComponent } from 'vue'
-import { DateLogic } from './dateLogic'
-
-export const MaskValue = defineComponent({
-	mixins: [DateLogic],
-	props: {
-		mask: {
-			type: [String, Boolean],
-			default: undefined,
-		},
-	},
-	computed: {
-		maskValue(): { mask: string } | undefined {
-			if (this.mask === false) {
-				return
-			}
-
-			if (typeof this.mask === 'string') {
-				return { mask: this.mask }
-			}
-
-			return {
-				mask: this.dateFormat.replace(/[a-z]/gim, '#'),
-			}
-		},
-	},
-})
