--- packages/synapse-bridge/src/patterns/DatePickerField/config.ts
@@ -1,17 +1,0 @@
-export const config = {
-	textField: {
-		hideDetails: 'auto',
-		persistentHint: true,
-		validateOn: 'blur',
-		color: '#0C419A',
-		maxErrors: 5,
-	},
-	datePicker: {
-		hideOffsetDates: true,
-		dayNames: ['L', 'M', 'M', 'J', 'V', 'S', 'D'],
-		locale: 'fr',
-	},
-	icon: {
-		color: '#808080',
-	},
-}
