--- packages/synapse-bridge/src/patterns/DatePickerField/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DatePickerField.vue'
